Separate scripts for readability and maintainability,
