+++ CartPage/product_cart.html
@@ -31,7 +31,7 @@
         </div>
 
         <div class="mt-4 text-center">
-            <a href="product_listing.html" class="btn btn-outline-dark">← Continue Shopping</a>
+            <a href="/ProductListing/product_listing.html" class="btn btn-outline-dark">← Continue Shopping</a>
         </div>
     </div>
