+++ PageDetails/prod_det_style.css
@@ -2,10 +2,10 @@ body {
     background-color: #f8f9fa;
     
 }
-/* 
+
 *{
     border: 1px dotted red;
-} */
+}
 
 /* Navbar profile image s*/
 .profile-img {
