+++ ProductListing/product_listing.html
@@ -29,7 +29,7 @@
 
     <section class="container mt-3 d-flex justify-content-between align-items-center">
         <input type="text" class="form-control w-75" id="search" placeholder="Search shoes...">
-        <a href="cart.html" class="btn btn-warning ms-3" id="cart-link">Cart (0)</a>
+        <a href="/CartPage/product_cart.html" class="btn btn-warning ms-3" id="cart-link">Cart (0)</a>
     </section>
 
     <main class="container mt-4">
@@ -90,8 +90,36 @@
         </div>
     </div>
     
-
+    <section id="testimonials" class="testimonials">
+        <h2 class="fw-bold text-center">WHAT OUR CUSTOMERS SAY</h2>
+        <div class="container mt-4">
+            <div class="row justify-content-center text-center">
+                <div class="col-md-3 d-flex align-items-stretch">
+                    <div class="p-3 text-white rounded d-flex flex-column justify-content-center">
+                        <p class="text-center">"Best shoes I've ever owned! The comfort and durability are unmatched. I wear them every day, and they still look brand new! The support they provide has made a huge difference in my daily routine. Whether I'm at work or out for a casual walk, these shoes never fail to keep me comfortable. Definitely worth every cent!" - Alcen</p>
+                    </div>
+                </div>
+                <div class="col-md-3 d-flex align-items-stretch">
+                    <div class="p-3 text-white rounded d-flex flex-column justify-content-center">
+                        <p class="text-center">"Comfort and style, all in one place. I’ve received so many compliments, and they fit perfectly! Worth every penny. The design is sleek, and I love how versatile they are—I can wear them to work, on a night out, or even during long walks without any discomfort. It’s rare to find shoes that combine fashion and function this well!" - Alfred</p>
+                    </div>
+                </div>
+                <div class="col-md-3 d-flex align-items-stretch">
+                    <div class="p-3 text-white rounded d-flex flex-column justify-content-center">
+                        <p class="text-center">"Exceptional quality, will buy again! The craftsmanship is top-notch, and I love the sleek design. Highly recommended! I’ve owned many brands before, but none have matched the level of detail and comfort these provide. Even after months of wear, they still look and feel great. These are my go-to shoes for any occasion" - Kenneth</p>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </section>
+    
+    <section id="contact" class="container contact-section text-center">
+        <p>123 Shoe Street, City, Country</p>
+        <p>Email: omsim@shoestore.com</p>
+        <p>Phone: +123 456 7890</p>
+    </section>
+    
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-    <script src="product_listing_script.js"></script>
+    <script type="module" src="./product_listing_script.js"></script>
 </body>
 </html>
