+++ ProductListing/product_listing_styles.css
@@ -69,9 +69,52 @@ body {
     transform: scale(1.25);
 }
 
+.price {
+    background-color: rgb(167, 167, 167);
+    color: white;
+    font-weight: bold;
+    padding: 8px 15px;
+    border-radius: 20px;
+    font-size: 16px;
+    display: inline-block;
+    text-align: center;
+    box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.2);
+}
+
 li,
 li a {
     list-style: none;
     text-decoration: none;
     color: black;
+}
+
+.testimonials {
+    background: black;
+    color: white;
+    padding: 60px 0;
+    text-align: center;
+    width: 100%;
+}
+
+.contact-section {
+    padding: 60px 0;
+    background: #f8f9fa;
+    width: 100%;
+
+}
+
+.contact-section p {
+    color: black;
+}
+
+.container {
+    width: 100%;
+    max-width: 100%;
+    padding-left: 0;
+    padding-right: 0;
+}
+
+.row {
+    margin-left: 0;
+    margin-right: 0;
 }
